project-01-ai-lead-classifier-openai/src/requirements.txt
# No external dependencies required

python-dotenv
openai
pandas

